+++ components/Search/components/Filter/Search.Filter.tsx
@@ -7,11 +7,11 @@ import {
 } from "@mui/material";
 import { FC, ReactNode } from "react";
 
-import { Maybe, Nullable } from "../../../../helpers/typings/utility-types";
 import {
   BucketsAggregations,
   ListBucketsAggregations,
-} from "../../../../lib/elastic/type";
+} from "../../../../core/elastic/type";
+import { Maybe, Nullable } from "../../../../helpers/typings/utility-types";
 import { useFilter } from "../../utils/hooks/useFilter";
 import classes from "./Search.filter.module.css";
 
@@ -36,41 +36,44 @@ export const SearchFilters: FC<ISearchFilters> = ({
     <div className={classes.root}>
       <FormControl fullWidth size="small">
         <FormLabel component="legend">{legend}</FormLabel>
-        <FormGroup className={classes.formGroup}>
-          {facetList.map((facet) => (
-            <FormControlLabel
-              key={facet.key}
-              // disableTypography
-              label={
-                <>
-                  <span className={classes.text}>{facet.key}</span>
-                  {nodes?.[facet.key] ? (
-                    <span className={classes.docCount}>
-                      {nodes && nodes[facet.key]
-                        ? nodes[facet.key]
-                        : facet.doc_count}
-                    </span>
-                  ) : (
-                    filter.list?.[facet.key] && (
-                      <span className={classes.docCount}>0</span>
-                    )
-                  )}
-                </>
-              }
-              control={
-                <Checkbox
-                  sx={{ padding: "4px 6px 4px 9px" }}
-                  checked={filter.list?.[facet.key] ?? false}
-                  disabled={
-                    !!(nodes && !nodes[facet.key]) && !filter.list?.[facet.key]
-                  }
-                  name={facet.key}
-                  // size="small"
-                  onChange={filter.hendleOnChange}
-                />
-              }
-            />
-          ))}
+        <FormGroup sx={{ display: "block" }} className={classes.formGroup}>
+          <div className={classes.body}>
+            {facetList.map((facet) => (
+              <FormControlLabel
+                key={facet.key}
+                // disableTypography
+                label={
+                  <>
+                    <span className={classes.text}>{facet.key}</span>
+                    {nodes?.[facet.key] ? (
+                      <span className={classes.docCount}>
+                        {nodes && nodes[facet.key]
+                          ? nodes[facet.key]
+                          : facet.doc_count}
+                      </span>
+                    ) : (
+                      filter.list?.[facet.key] && (
+                        <span className={classes.docCount}>0</span>
+                      )
+                    )}
+                  </>
+                }
+                control={
+                  <Checkbox
+                    sx={{ padding: "4px 6px 4px 9px" }}
+                    checked={filter.list?.[facet.key] ?? false}
+                    disabled={
+                      !!(nodes && !nodes[facet.key]) &&
+                      !filter.list?.[facet.key]
+                    }
+                    name={facet.key}
+                    // size="small"
+                    onChange={filter.hendleOnChange}
+                  />
+                }
+              />
+            ))}
+          </div>
         </FormGroup>
       </FormControl>
     </div>
