+++ components/Search/utils/index.ts
@@ -1,3 +1,2 @@
-export * from "../../../lib/elastic/convertArrayToObject";
 export * from "./fetchSearchData";
 export * from "./getFilterStringFromList";
