+++ core/elastic/constant/index.ts
@@ -1,0 +1,5 @@
+/**
+ * Search
+ */
+export const FACET_SIZE = 100;
+export const SEARCH_PATHNAME = "/search";
