+++ core/elastic/queryBlock/aggs.ts
@@ -1,9 +1,11 @@
 import { AggregationsAggregationContainer } from "@elastic/elasticsearch/api/types";
 
+import { FACET_SIZE } from "../../../constant/index";
+
 const departments: AggregationsAggregationContainer = {
   terms: {
     field: "departments.name.raw",
-    size: parseInt(process.env.FACET_SIZE ?? "10", 10),
+    size: FACET_SIZE,
   },
 };
 
@@ -11,7 +13,7 @@ const categories: AggregationsAggregationContainer = {
   terms: {
     field: "categories.name.raw",
     order: { _key: "asc" },
-    size: parseInt(process.env.FACET_SIZE ?? "10", 10),
+    size: FACET_SIZE,
   },
 };
