+++ core/elastic/queryBlock/highlight.ts
