+++ core/elastic/utility/mapping.json
