+++ lib/elastic/client.ts
