+++ pages/search/index.tsx
@@ -2,10 +2,10 @@ import { GetServerSideProps, NextPage } from "next";
 
 import Head from "../../components/Head/Head";
 import Layout from "../../components/UI/Layout/Layout";
+import { searchQuery } from "../../core/elastic/searchQuery";
 import { exceptionLog } from "../../helpers";
 import { getMenu, getUAPlatform } from "../../helpers/backend";
 import { UA } from "../../helpers/backend/getUA/const";
-import { searchQuery } from "../../lib/elastic/searchQuery";
 import { RouteSearch, RouteSearchProps } from "../../routes";
 
 type ISearchProps = {
