+++ routes/Search/Result/Search.ResultList.tsx
@@ -3,8 +3,8 @@ import { FC } from "react";
 
 import { Card, ICardProps } from "../../../components/Widget/Card/Card";
 import { Paragraph } from "../../../components/blocks/Paragraph/Paragraph";
+import { SearchHits } from "../../../core/elastic/type";
 import { Maybe, Omit } from "../../../helpers/typings/utility-types";
-import { SearchHits } from "../../../lib/elastic/type";
 
 type Data = { data: Maybe<SearchHits> };
 type ISearchResultListProps = Data & Omit<ICardProps, keyof Data>;
