+++ routes/Search/Route.Search.module.css
@@ -25,7 +25,7 @@
   .Aside {
     flex-basis: 280px;
     flex-shrink: 0;
-    padding: 10px 10px 0 0;
+    padding: 10px 20px 0 0;
   }
 }
 
@@ -33,6 +33,7 @@
   position: relative;
   display: flex;
   flex-wrap: wrap;
+  margin: 0 -10px;
 }
 @media screen and (width < 1100px) {
   .result {
